% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_id.R
\name{assign_id}
\alias{assign_id}
\title{Assign id to widgets}
\usage{
assign_id(sections)
}
\description{
Name the questions (ind_1 ... ind_n) - this slightly reduces the
tedious filling in of question numbers in .json
and reduces the likelihood of a manual mistake.
}
