% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_i18n.R
\name{df2list}
\alias{df2list}
\title{Transform dataframe to list}
\usage{
df2list(x)
}
\arguments{
\item{x}{a data frame with translations in rows and languages in columns}
}
\value{
a list of length 2 with x[[1]] has the vector of languages and x[[2]] has the translations
}
\description{
Transform dataframe to list
}
