% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_document.R
\name{render_document}
\alias{composeRmd}
\alias{composeSections}
\alias{composeQuestions}
\alias{renderReferences}
\title{Helper functions to generate the RMD file}
\usage{
composeRmd(
  answers = NULL,
  sectionsList = NULL,
  headList = NULL,
  answerList = NULL,
  language_code = NULL,
  save_as = "pdf"
)

composeSections(section, answers = NULL, language_code = NULL, save_as)

composeQuestions(question, answers = answers, language_code = NULL, save_as)

renderReferences(language_code = NULL)
}
\description{
These functions help to generate the output rmarkdown file containing the filled out
Transparency Checklist as a string without a template on the fly.
}
