% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_app.R
\name{renderSection}
\alias{renderSection}
\title{Functions to create the HTML structure of the shiny app dynamically}
\usage{
renderSection(section, id = NULL, answers = NULL)
}
\description{
These functions, first, read the header and the questions sections content and structure
from the appropriate internal datafile depending on whether a long or short checklist
is being built. Second, they create the HTML DOM tree dynamically when they called. To
save computing power we use them only once on app start.
}
\section{Warning}{
 The dynamically created HTML element ids should be namespaced
  in order to reach their inputs from the server side of the modules.
}

